Fix SSL handshake issues between pepd and pdp
The upgrade to OpenSAML 2.6.4 introduced a regression whereby
OpenSAML would override our HTTP client configuration, leading
to connection errors when TLS is enabled between the pepd and the pdp.

Luckily, OpenSAML provides a system property to disable this behaviour.
The system property is now defined in the PEPDaemon main.

The StrictHostnameVerifier is now configured explicitly for the HTTP
client.

Setting hostname verification triggers this JDK bug:

https://bugs.openjdk.java.net/browse/JDK-8133196

For which a workaround is to set the jdk.tls.trustNameService system
property to 'true' in PEPD_JAVA_OPTS.